test: verschiebe Upload- und LLM-Tests
